requirements.txt
rich>=13.0
click>=8.0
pyyaml>=6.0
rumps>=0.4
